Add files via upload
A web page fetching the weather forecast and allowing the users to take notes In this project, I have used the basic knowledge of JS, HTML, and CSS. It is to brush up on my learning outcomes such as fetching an API to display the weather forecast and some suggestions based on it. A note-making interface where I have used the concept of local storage to store previously written notes and remove them whenever the notes have been deleted. Also, have implemented CSS to create a cursor and make it move by using the coordinates of the cursor a bit of transition is also applied. during this mini-project, I got to know a lot about CSS I still have a lot to learn about CSS so all in all, it was just to test my abilities on a smaller scale.